+++ packages/react-components/react-button/src/stories/Button/ButtonWithLongText.stories.tsx
@@ -1,16 +1,27 @@
 import * as React from 'react';
-import { Button } from '@fluentui/react-components';
+import { makeStyles, Button } from '@fluentui/react-components';
 
-export const WithLongText = () => (
-  <>
-    <Button>Short text</Button>
-    <Button>Long text truncates after it hits the max width of the component</Button>
-  </>
-);
+const useStyles = makeStyles({
+  longText: {
+    width: '280px',
+  },
+});
+
+export const WithLongText = () => {
+  const styles = useStyles();
+
+  return (
+    <>
+      <Button>Short text</Button>
+      <Button className={styles.longText}>Long text wraps after it hits the max width of the component</Button>
+    </>
+  );
+};
+
 WithLongText.parameters = {
   docs: {
     description: {
-      story: 'Text truncates after it hits the max width of the component.',
+      story: 'Text wraps after it hits the max width of the component.',
     },
   },
 };