+++ packages/react-components/react-field/src/components/Field/Field.types.ts
@@ -24,9 +24,6 @@ export type FieldSlots<T extends FieldComponent> = {
 
   /**
    * The underlying component wrapped by this field.
-   *
-   * This is the PRIMARY slot: all intrinsic HTML properties will be applied to this slot,
-   * except `className` and `style`, which remain on the root slot.
    */
   control: SlotComponent<T>;
 
@@ -90,6 +87,32 @@ export type OptionalFieldComponentProps = {
    * Number sizes will be ignored, but are allowed because the HTML `<input>` element has a prop `size?: number`.
    */
   size?: 'small' | 'medium' | 'large' | number;
+};
+
+/**
+ * Configuration parameters for a Field class, passed to useField_unstable
+ */
+export type FieldConfig<T extends FieldComponent> = {
+  /**
+   * The underlying input component that this field is wrapping.
+   */
+  component: T;
+
+  /**
+   * Class names for this component, created by `getFieldClassNames`.
+   */
+  classNames: SlotClassNames<FieldSlots<T>>;
+
+  /**
+   * How the label be connected to the control.
+   * * htmlFor - Set the Label's htmlFor prop to the component's ID (and generate an ID if not provided).
+   *   This is the preferred method for components that use the underlying <input> tag.
+   * * aria-labelledby - Set the component's aria-labelledby prop to the Label's ID. Use this for components
+   *   that are not directly <input> elements (such as RadioGroup).
+   *
+   * @default htmlFor
+   */
+  labelConnection?: 'htmlFor' | 'aria-labelledby';
 };
 
 /**