+++ packages/react-components/react-field/src/components/TextareaField/TextareaField.test.tsx
@@ -1,0 +1,32 @@
+import * as React from 'react';
+import { resetIdsForTests } from '@fluentui/react-utilities';
+import { render } from '@testing-library/react';
+import { isConformant } from '../../common/isConformant';
+import { TextareaField } from './TextareaField';
+
+describe('TextareaField', () => {
+  isConformant({
+    Component: TextareaField,
+    displayName: 'TextareaField',
+    primarySlot: 'control',
+    testOptions: {
+      'has-static-classnames': [
+        {
+          props: {
+            label: 'label',
+            validationState: 'error',
+            validationMessage: 'validationMessage',
+            hint: 'hint',
+          },
+        },
+      ],
+    },
+  });
+
+  beforeEach(resetIdsForTests);
+
+  it('renders a default state', () => {
+    const result = render(<TextareaField />);
+    expect(result.container).toMatchSnapshot();
+  });
+});